Basic JS Keyboard input handling
This is a temporary port of the keyboard handling in the official Java
implementation to JS. It feels pretty laggy, but i hope this will change
after porting the stdlib to rust.

In the future, the keyboard implementation should also be ported to
rust. Not just for performance benefits, but also to make it easier to
build a native GUI for the VM independent from the browser.