Added desktop mode & tracing
this was done to find an annoying crash in my project 9 tetris game. But
after tracing the entire the vm for hours, I'm pretty sure, that it's
not a bug in the emulator, but actually a bug in my tetris
implementation. The reason it didn't show in the official emulator is due to their stdlib implementation being in java