Changed VM bytecode representation
Before this change, the bytecode was represented as actual bytes in
order to not waste any space inside the program vector to achieve better
cache usage.
After this change the bytecode is now represented in a way more
idiomatic enum. This simplifies the code in multiple places, but
especially in tests.
There is no noticable performance difference (but i should do a proper
benchmark in the future), so the old representation was probably a case
of premature optimization