Replaced keyboard.js with rust implementation
This has the advantage, that we can use the rust keyboard code in the
desktop version too. I don't think there's a performance benefit due to
the context switch between JS and rust though. It might even be slightly
slower this way, but it does remove some code duplication, which is
always nice.

This implementation has an intended bug due to compatibility issues with
the official emulator.
There is no way to input lowercase characters into the official
emulator, even though that would be trivial from a technical
perspective. A lot of games rely on this behaviour and expect all inputs
to be uppercase.
To keep compatibility with all of the existing games for the platform I
decided to capitalize all inputs on purpose in keyboard.rs