Added CodeView in Browser UI
The code view will only be displayed if the VM is not running. The
animation view in the original emulator is rarely used, but still uses a
lot of screen real estate. Wenn the VM is running, the CodeView will
automatically be hidden.

Also a lot of CSS styling. I did not include any special CSS for MS IE.
This browser will no longer be supported.
The UI is supposed to look minimalistic and kinda old-school on purpose
because it fits the theme of the project. It is however at least
partially adaptive. The CodeView will move below the Screen when the
view is less than 1000px wide